requirements.txt
streamlit==1.36.0
Flask==3.0.3
numpy==1.24.3
Pillow==10.4.0
gunicorn==22.0.0
torch==2.1.2
tensorflow-cpu==2.13.1
opencv-python==4.11.0.86
h5py==3.11.0
scikit-learn==1.3.2
matplotlib==3.7.5
torchvision==0.16.2
filelock
fsspec
networkx
sympy
mpmath
triton==2.1.0